refactor: manager들 불필요한 네이밍 수정
