+++ src/ts/types/CoinManager.ts
