+++ src/ts/types/ProductManager.ts
